Remove `rendering.rs` and refactor diff viewer to simplify rendering logic.
