+++ crates/diff_viewer/src/connector.rs
@@ -7,6 +7,7 @@ pub enum ConnectorKind {
 
 #[derive(Debug, Clone)]
 pub struct ConnectorCurve {
+    pub block_index: usize,
     pub focus_line: usize,
     pub left_start: usize,
     pub left_end: usize,
@@ -19,6 +20,7 @@ pub struct ConnectorCurve {
 
 impl ConnectorCurve {
     pub fn new(
+        block_index: usize,
         focus_line: usize,
         left_start: usize,
         left_end: usize,
@@ -29,6 +31,7 @@ impl ConnectorCurve {
         right_crushed: bool,
     ) -> Self {
         Self {
+            block_index,
             focus_line,
             left_start: left_start.min(left_end),
             left_end: left_end.max(left_start),
