+++ crates/diff_viewer/src/constants.rs
@@ -1,0 +1,24 @@
+/// Height in pixels for crushed blocks (empty ranges represented as thin lines)
+pub const CRUSHED_BLOCK_HEIGHT: f32 = 2.0;
+
+/// Width in pixels of the connector gutter between left and right editors
+pub const CONNECTOR_GUTTER_WIDTH: f32 = 45.0;
+
+/// Number of segments used to render Bezier curves for connectors
+pub const BEZIER_SEGMENTS: usize = 48;
+
+/// Alpha transparency value for diff highlight backgrounds
+pub const DIFF_HIGHLIGHT_ALPHA: f32 = 0.5;
+
+/// Base control point offset ratio for Bezier curves (relative to gutter width)
+pub const CONNECTOR_BASE_CONTROL_OFFSET_RATIO: f32 = 0.25;
+
+/// Thickness in pixels for crushed block indicator lines
+pub const CRUSHED_THICKNESS: f32 = 2.0;
+
+/// Default viewport height in pixels (fallback value)
+pub const DEFAULT_VIEWPORT_HEIGHT: f32 = 600.0;
+
+/// Default line height in pixels (fallback value)
+pub const DEFAULT_LINE_HEIGHT: f32 = 22.0;
+
