+++ crates/diff_viewer/src/lib.rs
@@ -1,5 +1,6 @@
 mod connector;
 mod connector_builder;
+mod constants;
 mod diff_operations;
 mod rendering;
 pub mod viewer;
