+++ crates/diff_viewer/src/rendering/mod.rs
@@ -1,0 +1,5 @@
+pub mod colors;
+pub mod connectors;
+pub mod crushed_blocks;
+pub mod revert_buttons;
+
